Improve split in backend convertors in tiger
Now the convertors are split in version folders.

JIRA: RAIL-2734
