+++ libs/sdk-backend-tiger/src/convertors/fromBackend/visualizationObjects/VisualizationObjectConverter.ts
@@ -1,0 +1,23 @@
+// (C) 2019-2021 GoodData Corporation
+import invariant from "ts-invariant";
+import { IInsightDefinition } from "@gooddata/sdk-model";
+import { VisualizationObjectModelV1, VisualizationObjectModelV2 } from "@gooddata/api-client-tiger";
+import { convertVisualizationObject as convertVisualizationObjectV1 } from "./v1/VisualizationObjectConverter";
+import { convertVisualizationObject as convertVisualizationObjectV2 } from "./v2/VisualizationObjectConverter";
+
+export const convertVisualizationObject = (
+    visualizationObject:
+        | VisualizationObjectModelV1.IVisualizationObject
+        | VisualizationObjectModelV2.IVisualizationObject,
+    title: string,
+): IInsightDefinition => {
+    if (VisualizationObjectModelV1.isVisualizationObject(visualizationObject)) {
+        return convertVisualizationObjectV1(visualizationObject);
+    }
+
+    if (VisualizationObjectModelV2.isVisualizationObject(visualizationObject)) {
+        return convertVisualizationObjectV2(visualizationObject, title);
+    }
+
+    invariant(false, "Unknown visualization object version");
+};
